Various refactor (#5870)
* various refactoring

* type

* fix avatar animation graph if controller component doesnt exist

* fix test and cleanup other tests
