+++ packages/engine/src/networking/functions/NetworkActionReceptor.test.ts
@@ -35,21 +35,17 @@ describe('NetworkActionReceptors', () => {
       assert.equal(world.userIdToUserIndex.get(userId), userIndex)
     })
 
-    it('should not add client if host', () => {
+    it('should not add client if already exists', () => {
       const world = Engine.instance.currentWorld
       const userId = 'user id' as UserId
       const userName = 'user name'
+      const userName2 = 'user name 2'
       const userIndex = 1
-      world.hostId = userId as UserId
-      Engine.instance.userId = userId
 
       NetworkActionReceptor.addClient(world, userId, userName, userIndex)
+      NetworkActionReceptor.addClient(world, userId, userName2, userIndex)
 
-      assert(!world.clients.get(userId))
-      assert(!world.userIndexToUserId.get(userIndex))
-      assert(!world.userIdToUserIndex.get(userId))
-
-      Engine.instance.userId = undefined!
+      assert(world.clients.get(userId)?.name, userName)
     })
   })
 
@@ -341,7 +337,6 @@ describe('NetworkActionReceptors', () => {
 
       const world = Engine.instance.currentWorld
       NetworkActionReceptor.addClient(world, userId, userName, userIndex)
-      Engine.instance.currentWorld = world
 
       const hostIndex = 0
       world.clients.set(world.hostId, { userId: world.hostId, name: 'server', index: hostIndex })
