+++ packages/engine/src/scene/functions/loaders/BoxColliderFunctions.test.ts
@@ -57,7 +57,6 @@ describe('BoxColliderFunctions', () => {
   })
 
   afterEach(() => {
-    Engine.instance.currentWorld = null!
     delete (globalThis as any).PhysX
   })
